// src/http/models/User.model.ts
export interface UserReference {
  id: string;
  name: string;
  username: string;
  profilePhoto: string;
  social_links: SocialLinks;
  joined: Date;
}

export interface IUser {
  id: string;
  name: string;
  username: string;
  email: string;
  profilePhoto: string;
  education: string;
  designation: string;
  bio: string;
  website_url: string;
  location: string;
  social_links: SocialLinks;
  profile_readme: string;
  skills: null;
  created_at: Date;
  updated_at: Date;
  two_factor_secret: null;
  two_factor_recovery_codes: null;
  profile_readme_html?: string;
}

export interface SocialLinks {
  github?: string;
  facebook?: string;
  stackOverflow?: string;
  medium?: string;
  linkedin?: string;
  twitter?: null;
  instagram?: string;
  behance?: string;
  dribbble?: string;
  twitch?: null;
  youtube?: string;
}